non auth file detail
