+++ packages/server/src/handler/file.ts
@@ -44,56 +44,6 @@ export const fileHandler = async (server: FastifyInstance) => {
     })
   })
 
-  server.get<{ Params: { fileId: string }; Reply: ResponseBody }>(
-    "/file/:fileId",
-    async (req, res) => {
-      const fileId = req.params.fileId
-
-      const repository = connection.getRepository(File)
-      const query = repository.createQueryBuilder("file")
-
-      const file = await query
-        .innerJoinAndSelect("file.author", "file_author")
-        .innerJoinAndSelect("file.updated_by", "file_updated_by")
-        .leftJoinAndSelect("file.stamps", "stamp")
-        .leftJoinAndSelect("stamp.author", "stamp_author")
-        .leftJoinAndSelect("stamp.comments", "comment")
-        .leftJoinAndSelect("comment.author", "comment_author")
-        .where("file.id = :fileId", { fileId })
-        .getOne()
-
-      if (!file) {
-        const e = createError(
-          ERR_BAD_URL,
-          `File(${fileId}) was not found.`,
-          404,
-        )
-        throw new e()
-      }
-
-      file.shared_to = [req.currentUser]
-      await repository.save(file)
-
-      const stamps = await file.stamps
-      const stampsResponse = await Promise.all(
-        stamps.map((s) => buildStampResponse(s)),
-      )
-
-      res.send({
-        result: "success",
-        data: {
-          fileSnapshot: {
-            type: file.fileType(req.currentUser),
-            file: buildFileResponse(file),
-            updatedAt: file.updated_at,
-            updatedBy: buildUserResponse(file.updated_by),
-          },
-          stamps: stampsResponse,
-        },
-      })
-    },
-  )
-
   server.post<{
     Body: { name: MultipartValue<string>; file: MultipartValue<string> }
     Reply: ResponseBody
